Partition chained assignment fixtures
Move the chained constant assignment compiler fixtures beside their sole owning test class without changing the fixture declarations or method bodies.
